message update

+++ src/main/java/com/keonah/stanbytest/controller/MemberController.java
@@ -21,11 +21,15 @@ public class MemberController {
 
     @PostMapping("")
     public ResponseEntity createMember(@Valid @RequestBody MemberCreateDto memberCreateDto) {
-        if(!memberService.checkNameAndTeamDuplication(memberCreateDto.getName(), memberCreateDto.getTeam()) && !memberRepository.existsByNo(memberCreateDto.getNo())) {
-            memberService.createMember(memberCreateDto);
-            return new ResponseEntity(memberCreateDto,HttpStatus.OK);
+        if(!memberService.checkNameAndTeamDuplication(memberCreateDto.getName(), memberCreateDto.getTeam())) {
+            if(!memberRepository.existsByNo(memberCreateDto.getNo())) {
+                memberService.createMember(memberCreateDto);
+                return new ResponseEntity(memberCreateDto, HttpStatus.OK);
+            } else {
+                return new ResponseEntity("사원 번호는 중복될 수 없습니다.", HttpStatus.CONFLICT);
+            }
         } else {
-            return new ResponseEntity(memberCreateDto,HttpStatus.CONFLICT);
+            return new ResponseEntity("한 부서에 같은 이름의 회원을 저장할 수 없습니다.", HttpStatus.CONFLICT);
         }
 
     }
@@ -36,16 +40,19 @@ public class MemberController {
             memberService.updateMember(no, memberUpdateDto);
             return new ResponseEntity(memberUpdateDto,HttpStatus.OK);
         } else {
-            return new ResponseEntity(memberUpdateDto,HttpStatus.CONFLICT);
+            return new ResponseEntity("한 부서에 같은 이름의 회원을 저장할 수 없습니다.", HttpStatus.CONFLICT);
         }
 
     }
 
     @DeleteMapping("/{no}")
     public ResponseEntity deleteMember(@PathVariable String no) {
+        if(memberRepository.findByNo(no) != null) {
             memberService.deleteMember(no);
-
-            return new ResponseEntity(no,HttpStatus.OK);
+            return new ResponseEntity("삭제 완료.", HttpStatus.OK);
+        } else {
+            return new ResponseEntity("존재하지 않는 회원입니다.", HttpStatus.BAD_REQUEST);
+        }
 
     }
